Refactor: rename ds component to search
Rename the ds (Deductive Search) component to search across the entire codebase:
- Renamed ddss/ds.py -> ddss/search.py
- Renamed ddss/ds.ts -> ddss/search.ts
- Renamed tests/test_ds.py -> tests/test_search.py
- Renamed tests/test_ds.ts -> tests/test_search.ts
- Updated all imports and references in main.py, main.ts
- Updated documentation in README.md, docs/en/*.md, docs/zh/*.md
- Updated test descriptions and function names

The dependency interface remains unchanged; only the component name is changed from ds to search.
