+++ docs/zh/modules.md
@@ -6,5 +6,5 @@
 - **Output** (`ddss/output.py`, `ddss/output.ts`)：实时显示数据库中的事实和想法。
 - **Load** (`ddss/load.py`, `ddss/load.ts`)：从标准输入批量导入事实。
 - **Dump** (`ddss/dump.py`, `ddss/dump.ts`)：将所有事实和想法导出到输出。
-- **DS** (`ddss/ds.py`, `ddss/ds.ts`)：前向链接演绎搜索引擎。
+- **Search** (`ddss/search.py`, `ddss/search.ts`)：前向链接演绎搜索引擎。
 - **Egg** (`ddss/egg.py`, `ddss/egg.ts`)：基于 E-graph 的等式推理引擎。
